fix(ci): apply 6 round-3 review fixes - int64_t timestamp_ns, OperationalError guard, DAB-11D test, checkout@v6, include/** filter
Agent-Logs-Url: https://github.com/foxrouter/meek/sessions/b090431b-437a-4894-ad66-376b5cd2c7e1

+++ src/main.cpp
@@ -438,10 +438,9 @@ static void capture_loop(std::stop_token st, ISdrSource& sdr,
     }
 
     blk.samples.assign(buf.begin(), buf.begin() + n);
-    blk.timestamp_ns =
-        static_cast<std::uint64_t>(std::chrono::duration_cast<std::chrono::nanoseconds>(
-                                       std::chrono::system_clock::now().time_since_epoch())
-                                       .count());
+    blk.timestamp_ns = std::chrono::duration_cast<std::chrono::nanoseconds>(
+                           std::chrono::system_clock::now().time_since_epoch())
+                           .count();
     blk.center_freq_hz = sdr.center_freq_hz();
     blk.sample_rate_hz = sdr.sample_rate_hz();
